+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 3.9.1
 
 - Improvement: Added `TODO` comments on `NotImplementedException`.
+- Fix: Unique Constraint validation ignores Included columns.
 
 ### Version 3.9.0