+++ Modules/Intent.Modules.Application.MediatR.FluentValidation/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 4.6.1
 
 - Improvement: Added `TODO` comments on `NotImplementedException`.
+- Fix: Unique Constraint validation ignores Included columns.
 
 ### Version 4.6.0