+++ Modules/Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 1.1.1
 
 - Improvement: Added `TODO` comments on `NotImplementedException`.
+- Fix: Unique Constraint validation ignores Included columns.
 
 ### Version 1.1.0